+++ internal/config/config.go
@@ -6,7 +6,7 @@ import (
 )
 
 // Version is set at build time via ldflags.
-var Version = "0.12.2"
+var Version = "0.12.3"
 
 type Config struct {
 	MongoDBURI           string